Add proper calculation of conversion factor by load cell rating
This means that by inputting the relevant parameters for a given
load cell you can get force output in true kg, allowing us to swap
out for a different load cell while preserving the force curve.

The values given preserve the existing behaviour by generating the
same conversion factor, but this is likely to be inaccurate for
most users.